refactor utils to package to prepare for metadata utils
