+++ shift_detector/checks/statistical_checks/NumericalStatisticalCheck.py
@@ -1,8 +1,8 @@
 import pandas as pd
 from scipy import stats
 
-from shift_detector.Utils import ColumnType
 from shift_detector.checks.statistical_checks.StatisticalCheck import SimpleStatisticalCheck
+from shift_detector.utils.ColumnManagement import ColumnType
 
 
 def kolmogorov_smirnov_test(part1: pd.Series, part2: pd.Series):
