+++ shift_detector/preprocessors/DummyPreprocessor.py
@@ -1,8 +1,8 @@
 from pandas import DataFrame
 
-from shift_detector.Utils import ColumnType
 from shift_detector.preprocessors.Preprocessor import Preprocessor
 from shift_detector.preprocessors.Store import Store
+from shift_detector.utils.ColumnManagement import ColumnType
 
 
 class DummyPreprocessor(Preprocessor):
